refactor: 스크리닝 정보 페이지 생성
