+++ src/components/pages/detailInfoPage/DetailInfoPage.style.ts
