+++ src/components/pages/detailInfoPage/components/detailContact/DetailContact.style.ts
