+++ src/components/pages/detailInfoPage/components/detailContact/index.tsx
@@ -4,14 +4,15 @@ import Call from '@/assets/icons/call.svg';
 import Mail from '@/assets/icons/mail.svg';
 import palette from '@/styles/theme/color';
 import Typography from '@/components/typography';
-import {infoPhoneStyles} from './DetailInfoPhone.style';
+
+import {infoPhoneStyles} from './DetailContact.style';
 
 interface IDetailInfoPhoneProps {
   type: 'call' | 'mail';
   content: string;
 }
 
-const DetailInfoPhone = ({type, content}: IDetailInfoPhoneProps) => {
+const DetailContact = ({type, content}: IDetailInfoPhoneProps) => {
   return (
     <View style={infoPhoneStyles.content}>
       {type === 'call' && <Call />}
@@ -22,4 +23,4 @@ const DetailInfoPhone = ({type, content}: IDetailInfoPhoneProps) => {
     </View>
   );
 };
-export default DetailInfoPhone;
+export default DetailContact;
