+++ src/components/pages/detailInfoPage/components/detailInfo/index.tsx
