+++ src/screens/screening/detail/DetailScreen.tsx
@@ -5,8 +5,8 @@ import {useQuery} from '@tanstack/react-query';
 import DetailTitle from './components/detailTitle';
 import ImageContentScrollContainer from '@/components/container/imageContentScrollContainer';
 import TabBar from '@/components/tabBar';
+import DetailInfoPage from '@/components/pages/detailInfoPage';
 import DetailReviewScreen from './tabs/detailReviewScreen/DetailReviewScreen';
-import DetailInfoScreen from './tabs/detailInfoScreen/DetailInfoScreen';
 import stackScreens from '@/constants/stackScreens';
 import {ScreenRouteProp} from '@/types/navigator';
 import BottomDetailButton from './components/bottomDetailButton';
@@ -115,7 +115,7 @@ const DetailScreen = ({route}: DetailScreenProps) => {
 
             <View>
               {currentTab === 0 && data?.data && (
-                <DetailInfoScreen item={data?.data} />
+                <DetailInfoPage item={data?.data} />
               )}
               {currentTab === 1 && <DetailReviewScreen id={id} />}
             </View>
